Introduce `Hanami::API::DSL` which gives the ability to other frameworks to use the `Hanami::API` DSL
